Updated to latest upstream firmware and customized configuration for Sidewinder X1
I forked delwinbest's repo, and updated the build with the latest from BTT's upstream repo.

Looking at delwinbest's repo, there was no actual code customization necessary. It turns out that the `MKS_32_V1_4` build option from BTT works fine as-is along with the TFT28 images and fonts as long as the files are renamed and placed on the SD card where the MKS screen expects them to be.

I branched off of the latest code in the master branch of the upstream BTT repo, pulled in the `Copy to SD Card - MKS` folder from delwinbest's `bibo2_touch` branch, updated the `config.ini` file to match the settings for my Sidewinder X1 (you may to change the `baudrate` setting depending on your Marlin build, but the rest should work as-is on any Sidewinder), updated the bmp images to the latest upstream, added the custom Sidewinder `Logo.bmp` file from the zip file available for download at `https://3dprintbeginner.com/custom-screen-firmware-for-sidewinder-x1-and-genius`, rebuilt the firmware based on the latest upstream code, and updated the `mkstft23.bin` with that build.

Tested working perfectly so far on my Sidewinder X1 with stock control board and screen. It's way better than the stock screen firmware!